% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_form_create_rating.R
\name{peer_form_create_rating}
\alias{peer_form_create_rating}
\title{Create author rating form}
\usage{
peer_form_create_rating(
  category = c("helpfulness", "accuracy", "fairness"),
  title = "Author rating form",
  fname = "feedback_blank_rating",
  output = "github_document",
  write_rmd = TRUE,
  overwrite = FALSE,
  allow_comment = FALSE
)
}
\arguments{
\item{category}{Character. Categories to be included in the feedback form, defaults to \code{c("helpfulness", "accuracy", "fairness")}.}

\item{title}{Character. Title of form, defaults to "Author rating form".}

\item{fname}{Character. File name of RMarkdown document to be written to memory, defaults to \code{feedback_blank_rating}.}

\item{output}{Character. Output parameter for \code{.Rmd} file, defaults to \code{github_document}.}

\item{write_rmd}{Logical. Whether the feedback form should be saved to a \code{.Rmd} file in the current working directory, defaults to TRUE.}

\item{overwrite}{Logical. Should existing file or files with same name be overwritten, defaults to \code{FALSE}.}

\item{allow_comment}{Logical. Should optional comment field be included? Defaults to \code{FALSE}.}
}
\description{
\code{peer_form_create_rating} creates a short feedback form for authors to rate the feedback they got from reviewers. The rating categories are based on Reily, K. and P. Ludford Finnerty,  and L. Terveen (2009): Two Peers Are Better Than One: Aggregating Peer Reviews for Computing Assignments is Surprisingly Accurate. In \emph{Proceedings of the ACM 2009 International Conference on Supporting Group Work}. GROUP’09, May 10–13, 2009, Sanibel Island, Florida, USA.
}
\examples{
\dontrun{
peer_form_create_rating(c("accuracy", "fairness"))
}

}
\seealso{
Other peer review functions:
\code{\link[=peer_assign]{peer_assign()}},
\code{\link[=peer_file_add_aut]{peer_file_add_aut()}},
\code{\link[=peer_file_add_rev]{peer_file_add_rev()}},
\code{\link[=peer_form_create_review]{peer_form_create_review()}},
\code{\link[=peer_init]{peer_init()}},
\code{\link[=peer_return]{peer_return()}},
\code{\link[=peer_roster_create]{peer_roster_create()}},
\code{\link[=peer_score_rating]{peer_score_rating()}},
\code{\link[=peer_score_review]{peer_score_review()}}
}
\concept{peer review functions}
