% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_roster_create.R
\name{peer_roster_create}
\alias{peer_roster_create}
\title{Create peer review roster}
\usage{
peer_roster_create(n_rev, user, seed = NULL, write_csv = TRUE, dir = NULL)
}
\arguments{
\item{n_rev}{Numeric. Number of reviews per user. Must be larger than zero and smaller
than the number of users.}

\item{user}{Character. A vector of GitHub user names.}

\item{seed}{Numeric. Random seed for assignment, if not set a random seed will be chosen.}

\item{write_csv}{Logical. Whether the roster data frame should be written to a \code{.csv} file
in the current working directory, defaults to TRUE.}

\item{dir}{Character. Directory where the peer review roster will be written if \code{write_csv = TRUE}.}
}
\description{
\code{peer_roster_create} creates data frame of random assignments of author files to
reviewers. By default, the output is saved to a \code{.csv} file in the current working
directory that incorporates the current date and random seed as part of the file name.
}
\examples{
\dontrun{
peer_roster_create(
  3,
  c("anya-ghclass", "bruno-ghclass",
    "celine-ghclass", "diego-ghclass"),
  dir = "hw2/"
)
}

}
\seealso{
Other peer review functions:
\code{\link[=peer_assign]{peer_assign()}},
\code{\link[=peer_file_add_aut]{peer_file_add_aut()}},
\code{\link[=peer_file_add_rev]{peer_file_add_rev()}},
\code{\link[=peer_form_create_rating]{peer_form_create_rating()}},
\code{\link[=peer_form_create_review]{peer_form_create_review()}},
\code{\link[=peer_init]{peer_init()}},
\code{\link[=peer_return]{peer_return()}},
\code{\link[=peer_score_rating]{peer_score_rating()}},
\code{\link[=peer_score_review]{peer_score_review()}}
}
\concept{peer review functions}
