% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peer_file_add_rev.R
\name{peer_file_add_rev}
\alias{peer_file_add_rev}
\title{Add local files to author-specific folders on reviewers' review repositories}
\usage{
peer_file_add_rev(
  org,
  roster,
  local_path,
  prefix = "",
  suffix = "",
  message = NULL,
  branch = NULL,
  overwrite = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{org}{Character. Name of the GitHub organization.}

\item{roster}{Character. Data frame or file path of roster file with author-reviewer assignments. Must contain a column \code{user} with GitHub user names of authors, a column \code{user_random} with randomized tokens for user names, and one or more \verb{rev*} columns that specify review assignments as values of the vector \code{user_random}.}

\item{local_path}{Character. File name of file to be added.}

\item{prefix}{Character. Common repository name prefix.}

\item{suffix}{Character. Common repository name suffix.}

\item{message}{Character. Commit message.}

\item{branch}{Character. Name of branch the file should be committed to. If \code{NULL} (the default), the repository's default branch is used.}

\item{overwrite}{Logical. Whether existing files in reviewers' repositories should be overwritten, defaults to \code{FALSE}.}

\item{verbose}{Logical. Should success/failure messages be printed, defaults to \code{TRUE}.}
}
\description{
\code{peer_file_add_rev()} takes a local file and adds it to author-specific folders on reviewers' repositories. The function's main purpose is to distribute review forms into the correct author-specific folders on reviewers' repositories.
}
\examples{
\dontrun{
peer_file_add_rev(org = "ghclass-test",
roster = "hw2_roster_seed12345.csv",
local_path = "/Users/profx/introstats/hw2/hw2_review.Rmd",
prefix = prefix)
}

}
\seealso{
Other peer review functions:
\code{\link[=peer_assign]{peer_assign()}},
\code{\link[=peer_file_add_aut]{peer_file_add_aut()}},
\code{\link[=peer_form_create_rating]{peer_form_create_rating()}},
\code{\link[=peer_form_create_review]{peer_form_create_review()}},
\code{\link[=peer_init]{peer_init()}},
\code{\link[=peer_return]{peer_return()}},
\code{\link[=peer_roster_create]{peer_roster_create()}},
\code{\link[=peer_score_rating]{peer_score_rating()}},
\code{\link[=peer_score_review]{peer_score_review()}}
}
\concept{peer review functions}
